Feature/1499 (#1515)
* turing text url to links

* Update client/src/components/Feed/Linkify.js


* fix according to review

* used existing linkify to convert text links into html links

* removed regexify-string libarary

removed regexify-string libarary as it is no more used.
